feat: add index_path config option
This allows the user to pick where the output index file will be located
and called. I still want to add an option that allows the user to remove
categories if they want. This will take more work than I can afford
right now so I'll focus on individual repository pages for now. I'll
keep it in mind though!